+++ app/app.js
@@ -13,6 +13,7 @@ app.use(helmet.frameguard({ action: 'deny' }));
 app.use(helmet.noSniff());
 app.use(helmet.ieNoOpen());
 app.use(helmet.xssFilter());
+app.use(helmet.dnsPrefetchControl());
 app.use(helmet.referrerPolicy({ policy: 'no-referrer' }));
 app.use(helmet.hsts({
       maxAge: 31536000000,